Consolidate HTTPClient implementation into single file
